Move LoggerFactory to core
