+++ src/UA.ts
@@ -11,6 +11,9 @@ import {
   IncomingRequestMessage,
   IncomingResponseMessage,
   IncomingSubscribeRequest,
+  Levels,
+  Logger,
+  LoggerFactory,
   Transport,
   URI,
   UserAgentCore,
@@ -20,7 +23,6 @@ import {
 import { SessionStatus, TypeStrings, UAStatus } from "./Enums";
 import { Exceptions } from "./Exceptions";
 import { Grammar } from "./Grammar";
-import { Levels, Logger, LoggerFactory } from "./LoggerFactory";
 import { Parser } from "./Parser";
 import { PublishContext } from "./PublishContext";
 import { ReferServerContext } from "./ReferContext";
