+++ src/Web/SessionDescriptionHandler.ts
@@ -1,8 +1,8 @@
 import { EventEmitter } from "events";
 
+import { Logger } from "../core";
 import { TypeStrings } from "../Enums";
 import { Exceptions } from "../Exceptions";
-import { Logger } from "../LoggerFactory";
 import { InviteClientContext, InviteServerContext } from "../Session";
 import {
   BodyObj,
