+++ src/Web/Transport.ts
@@ -1,8 +1,7 @@
-import { OutgoingRequestMessage, Transport as TransportBase } from "../core";
+import { Logger, OutgoingRequestMessage, Transport as TransportBase } from "../core";
 import { TypeStrings } from "../Enums";
 import { Exceptions } from "../Exceptions";
 import { Grammar } from "../Grammar";
-import { Logger } from "../LoggerFactory";
 import { Utils } from "../Utils";
 
 export enum TransportStatus {
