+++ src/core/log/levels.ts
@@ -1,0 +1,10 @@
+/**
+ * Log levels.
+ * @public
+ */
+export enum Levels {
+  error,
+  warn,
+  log,
+  debug
+}
