+++ src/core/messages/digest-authentication.ts
@@ -1,6 +1,6 @@
 import MD5 from "crypto-js/md5";
 
-import { Logger, LoggerFactory } from "../../LoggerFactory";
+import { Logger, LoggerFactory } from "../log";
 import { OutgoingRequestMessage } from "./outgoing-request-message";
 import { URI } from "./uri";
 import { createRandomToken } from "./utils";
